Moved logic into library, moved server unit test into integration
In order to support integration testing, I've created a library that the main executable will reference.
Ive moded the server unit tests into a more comprehensive integration tests. In the future, I'd like to refactor the server
to support easier additon of unit tests, but the pasing of &TcpStream down throw all levels of the stack is proving difficult to mock and test.